refactor: move resolveConnectorId to ids-core
